+++ tests/aggregated_stream/test_stream_projection.py
@@ -1,10 +1,16 @@
 import threading as _threading
 import uuid as _uuid
+from unittest import mock as _mock
 
 import pytest
 import sqlalchemy as _sa
 
-from depeche_db._aggregated_stream import AggregatedStream, SelectedOriginStream
+from depeche_db._aggregated_stream import (
+    AggregatedStream,
+    FullUpdateResult,
+    SelectedOriginStream,
+)
+from depeche_db._interfaces import FixedTimeBudget, RunOnNotificationResult
 from tests._account_example import Account, AccountRepository
 
 
@@ -12,26 +18,26 @@ def test_stream_projector(db_engine, store_factory, stream_factory, account_ids)
     ACCOUNT1_ID, ACCOUNT2_ID = account_ids
     store = store_factory()
     subject = stream_factory(store)
-    assert subject.projector.update_full() == 0
+    assert subject.projector.update_full() == FullUpdateResult(0, False)
 
     account_repo = AccountRepository(store)
 
     account = Account.register(id=ACCOUNT1_ID, owner_id=_uuid.uuid4(), number="123")
     account.credit(100)
     account_repo.save(account, expected_version=0)
-    assert subject.projector.update_full() == 2
+    assert subject.projector.update_full() == FullUpdateResult(2, False)
 
     account2 = Account.register(id=ACCOUNT2_ID, owner_id=_uuid.uuid4(), number="234")
     account2.credit(100)
     account_repo.save(account2, expected_version=0)
-    assert subject.projector.update_full() == 2
+    assert subject.projector.update_full() == FullUpdateResult(2, False)
 
     account2.credit(100)
     account2.credit(100)
     account2.credit(100)
     account2.credit(100)
     account_repo.save(account2, expected_version=2)
-    assert subject.projector.update_full() == 4
+    assert subject.projector.update_full() == FullUpdateResult(4, False)
 
     assert_stream_projection(subject, db_engine, account, account2)
 
@@ -48,7 +54,7 @@ def test_stream_projector_origin_selection(
 
     def get_select_origin_streams():
         with db_engine.connect() as conn:
-            return subject.projector._select_origin_streams(conn=conn, cutoff_cond=[])
+            return subject.projector._select_origin_streams(conn=conn)
 
     account = Account.register(id=ACCOUNT1_ID, owner_id=_uuid.uuid4(), number="123")
     account.credit(100)
@@ -56,7 +62,7 @@ def test_stream_projector_origin_selection(
     account_repo.save(account, expected_version=0)
 
     assert get_select_origin_streams() == [
-        SelectedOriginStream(f"account-{account.id}", 0, 1, 3)
+        SelectedOriginStream(stream=f"account-{account.id}", start_at_global_position=1)
     ]
 
     account2 = Account.register(id=ACCOUNT2_ID, owner_id=_uuid.uuid4(), number="234")
@@ -65,11 +71,26 @@ def test_stream_projector_origin_selection(
     account2.credit(100)
     account_repo.save(account2, expected_version=0)
 
+    # because of the cached origin streams, the projector will not update
+    # right away...
     assert get_select_origin_streams() == [
-        SelectedOriginStream(f"account-{account.id}", 0, 1, 3),
-        SelectedOriginStream(f"account-{account2.id}", 0, 4, 4),
+        SelectedOriginStream(
+            stream=f"account-{account.id}", start_at_global_position=1
+        ),
     ]
-    assert subject.projector.update_full() == 7
+
+    # ... but only when the cache is cleared
+    subject.projector._get_origin_stream_positions_cache = None
+    assert get_select_origin_streams() == [
+        SelectedOriginStream(
+            stream=f"account-{account.id}", start_at_global_position=1
+        ),
+        SelectedOriginStream(
+            stream=f"account-{account2.id}", start_at_global_position=4
+        ),
+    ]
+
+    assert subject.projector.update_full() == FullUpdateResult(7, False)
 
 
 def test_stream_projector_origin_selection_late_commit(
@@ -83,7 +104,7 @@ def test_stream_projector_origin_selection_late_commit(
 
     def get_select_origin_streams():
         with db_engine.connect() as conn:
-            return subject.projector._select_origin_streams(conn=conn, cutoff_cond=[])
+            return subject.projector._select_origin_streams(conn=conn)
 
     with db_engine.connect() as long_running_conn:
         # Simulate a long-running transaction that commits later
@@ -104,20 +125,23 @@ def test_stream_projector_origin_selection_late_commit(
         )
         account2.credit(100)
         account_repo.save(account2, expected_version=0)
+        print("HERE")
         assert get_select_origin_streams() == [
             # min_global_position == 3 here because the events above already
             # consumed sequence numbers 1 and 2 even though they are not visible yet
-            SelectedOriginStream(f"account-{account2.id}", 0, 3, 2),
+            SelectedOriginStream(
+                stream=f"account-{account2.id}", start_at_global_position=3
+            ),
         ]
-        assert subject.projector.update_full() == 2
+        assert subject.projector.update_full() == FullUpdateResult(2, False)
         long_running_conn.commit()
 
     assert get_select_origin_streams() == [
         # min_global_position == 1 because the events from the long-running
         # transaction are now visible
-        SelectedOriginStream(f"account-{account.id}", 0, 1, 2),
+        SelectedOriginStream(f"account-{account.id}", start_at_global_position=1),
     ]
-    assert subject.projector.update_full() == 2
+    assert subject.projector.update_full() == FullUpdateResult(2, False)
     assert_stream_projection(subject, db_engine, account, account2)
 
 
@@ -160,6 +184,12 @@ def test_stream_projector_cutoff(db_engine, store_factory, stream_factory, accou
     ]
 
 
+def test_stream_projector_interest(store_factory, stream_factory):
+    subject: AggregatedStream = stream_factory(store_factory())
+    assert subject.projector.interested_in_notification({"stream": "account-123"})
+    assert not subject.projector.interested_in_notification({"stream": "foo"})
+
+
 def test_stream_projector_locking(db_engine, store_factory, stream_factory):
     subject = stream_factory(store_factory())
     with db_engine.connect() as conn:
@@ -193,3 +223,52 @@ def test_only_positive_partitions(
     subject = stream_factory(store, partitioner=IllegalPartitioner())
     with pytest.raises(ValueError):
         subject.projector.update_full()
+
+
+def test_stream_projector_run(db_engine, store_factory, stream_factory, account_ids):
+    ACCOUNT1_ID, _ = account_ids
+    store = store_factory()
+    subject = stream_factory(store, batch_size=2)
+    assert (
+        subject.projector.run(FixedTimeBudget(-1))
+        == RunOnNotificationResult.DONE_FOR_NOW
+    )
+
+    account_repo = AccountRepository(store)
+
+    account = Account.register(id=ACCOUNT1_ID, owner_id=_uuid.uuid4(), number="123")
+    account.credit(100)
+    account_repo.save(account, expected_version=0)
+    assert (
+        subject.projector.run(FixedTimeBudget(-1))
+        == RunOnNotificationResult.WORK_REMAINING
+    )
+
+
+def test_stream_projector_work_left(
+    db_engine, store_factory, stream_factory, account_ids
+):
+    ACCOUNT1_ID, _ = account_ids
+    store = store_factory()
+    subject = stream_factory(store, batch_size=2)
+    with _mock.patch.object(
+        subject.projector, "_update_batch", wraps=subject.projector._update_batch
+    ) as mock_update_full:
+        account_repo = AccountRepository(store)
+
+        account = Account.register(id=ACCOUNT1_ID, owner_id=_uuid.uuid4(), number="123")
+        account.credit(100)
+        account.credit(100)
+        account_repo.save(account, expected_version=0)
+
+        # This will only do one batch because of the batch size and the already exhausted time budget
+        assert subject.projector.update_full(FixedTimeBudget(-1)) == FullUpdateResult(
+            2, True
+        )
+
+        # This will do the rest of the batches
+        assert subject.projector.update_full() == FullUpdateResult(1, False)
+
+        # There are only 3 messages in total, so the total number of calls
+        # to _update_batch should be 2.
+        assert mock_update_full.call_count == 2