+++ docs/CHANGELOG.md
@@ -1,3 +1,15 @@
+# 0.11.0 (only release candidate)
+
+* Improved performance of aggregated stream projection.
+* Fixed bug in return of AggregatedStream.projector.run which would always re-run it in the executor right away.
+* **BREAKING CHANGE**: Extended the RunOnNotification protocol
+    * Added `interested_in_notification(dict) -> bool`
+    * Added `take_notification_hint(dict) -> bool`
+* **BREAKING CHANGE**: Changed DB schema of aggregated streams
+    * Added columns; you will have to do a downtime deployment for these changes
+    * Use the following command to generate a migration script:
+      `python -m depeche_db generate-migration-script <PREV-VERSION> 0.11 --message-store=<NAME> --aggregated-stream=<STREAM-NAME> --aggregated-stream=<ANOTHER...>`
+
 # 0.10.3
 
 * Use one insert operation per batch in aggregated stream update instead of one per message